+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Webhooks (beta) Changelog
 
+## 19th August 2024
+
+- Added proof archived and unarchived webhooks.
+- Added team id, due date and previous version proof id to proof payload.
+
 ## 2nd April 2024
 
 - Added proof overdue webhook.